backend/requirements.txt
fastapi[standard]>=0.109.2
pydantic>=2.6.1
pymongo>=4.6.1
motor>=3.3.2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
